requirements.txt
absl-py==2.1.0
accelerate==0.29.3
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.35.0
botocore==1.35.0
cachetools==5.4.0
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.7
cohere==5.8.1
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dlib==19.24.4
dnspython==2.6.1
docstring-to-markdown==0.15
ecdsa==0.19.0
einops==0.7.0
entrypoints==0.4
environs==9.5.0
executing==2.0.1
fastapi==0.105.0
fastavro==1.9.5
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2024.2.0
google==3.0.0
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gotrue==2.6.0
grpcio==1.65.4
grpcio-tools==1.65.4
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.22.2
hyperframe==6.0.1
idna==3.4
imageio==2.34.0
ipykernel==6.29.2
ipython==8.21.0
isoduration==20.11.0
jax==0.4.26
jaxlib>=0.4.34
jedi==0.19.1
jedi-language-server==0.41.4
Jinja2==3.1.3
jiter==0.6.1
jmespath==1.0.1
joblib==1.3.2
json5==0.9.17
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.3
jupyter_client==7.4.9
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
kiwisolver==1.4.5
langchain==0.2.12
langchain-cohere==0.2.2
langchain-community==0.2.11
langchain-core==0.2.28
langchain-experimental==0.0.64
langchain-openai==0.1.20
langchain-qdrant==0.1.3
langchain-text-splitters==0.2.2
langsmith==0.1.96
lazy_loader==0.4
lime==0.2.0.1
lmdb==1.4.1
lsprotocol==2023.0.1
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
#mediapipe==0.10.14
milvus-lite==2.4.8
mistune==3.0.2
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==6.0.4
multimetric==2.2.1
mypy==1.10.0
mypy-extensions==1.0.0
nbclient==0.9.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.9.1
notebook_shim==0.2.4
numpy
oauthlib==3.2.2
openai==1.51.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.10.6
overrides==7.7.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.3
pexpect==4.9.0
pika==1.3.2
Pillow==10.1.0
platformdirs==4.2.0
plotly==5.22.0
portalocker==2.10.1
postgrest==0.16.9
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydotenv==0.0.7
pygls==1.3.1
Pygments==2.17.2
pygraphviz==1.13
pylspclient==0.1.2
pymongo==4.7.3
pyparsing==3.1.1
pyright==1.1.365
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.16
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
qdrant-client==1.10.1
realtime==1.0.6
redis==5.0.7
referencing==0.33.0
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
scikit-image==0.23.1
scikit-learn==1.3.2
scipy==1.11.3
Send2Trash==1.8.2
six==1.16.0
smalldiffusion==0.3
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
sparseqr==1.2.1
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.27.0
storage3==0.7.7
StrEnum==0.4.15
supabase==2.5.3
supafunc==0.4.7
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2024.4.18
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.20.0
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.65.0
traitlets==5.14.1
tree-sitter==0.22.3
tree-sitter-builds==2022.8.27
tree-sitter-go==0.21.0
tree-sitter-java==0.21.0
tree-sitter-typescript==0.21.2
types-python-dateutil==2.8.19.20240106
types-requests==2.32.0.20240712
typesense==0.21.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.2
uvicorn==0.24.0.post1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
xxhash==3.4.1
yarl==1.9.2
